feedback-system-backend/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Werkzeug==3.1.3
python-dotenv==1.0.0
Flask-Migrate==4.1.0
gunicorn==21.2.0
SQLAlchemy==2.0.42
alembic==1.13.1
psycopg2-binary==2.9.7